requirements.txt
adjustText==0.7.3
alembic==1.6.5
anyio==3.2.1
argon2-cffi==20.1.0
arviz==0.11.2
async-generator==1.10
attrs==21.2.0
Babel==2.9.1
backcall==0.2.0
bioinfokit==2.0.4
bleach==3.3.1
cachetools==4.2.2
certifi==2021.5.30
cffi==1.14.6
cftime==1.5.0
charset-normalizer==2.0.3
cliff==3.8.0
cloudpickle==1.6.0
cmd2==2.1.2
colorama==0.4.4
colorlog==5.0.1
cycler==0.10.0
dask==2021.7.0
debugpy==1.3.0
decorator==5.0.9
defusedxml==0.7.1
dill==0.3.4
einops==0.3.0
entrypoints==0.3
fastprogress==1.0.0
filelock==3.0.12
fsspec==2021.7.0
greenlet==1.1.0
idna==3.2
ipykernel==6.0.3
ipython==7.25.0
ipython-genutils==0.2.0
ipywidgets==7.6.3
jedi==0.18.0
Jinja2==3.0.1
joblib==1.0.1
json5==0.9.6
jsonschema==3.2.0
jupyter-client==6.1.12
jupyter-core==4.7.1
jupyter-server==1.9.0
jupyterlab==3.0.16
jupyterlab-pygments==0.1.2
jupyterlab-server==2.6.1
jupyterlab-widgets==1.0.0
kiwisolver==1.3.1
locket==0.2.1
Mako==1.1.4
MarkupSafe==2.0.1
matplotlib==3.4.2
matplotlib-inline==0.1.2
matplotlib-venn==0.11.6
missingno==0.5.0
mistune==0.8.4
nbclassic==0.3.1
nbclient==0.5.3
nbconvert==6.1.0
nbformat==5.1.3
nest-asyncio==1.5.1
netCDF4==1.5.7
notebook==6.4.0
numpy==1.19.5
oauthlib==3.1.1
packaging==21.0
pandas==1.3.1
pandocfilters==1.4.3
parso==0.8.2
partd==1.2.0
patsy==0.5.1
pbr==5.6.0
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.3.1
plotly==5.1.0
prettytable==2.1.0
prometheus-client==0.11.0
prompt-toolkit==3.0.19
ptyprocess==0.7.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
Pygments==2.9.0
pyparsing==2.4.7
pyperclip==1.8.2
pyrsistent==0.18.0
python-dateutil==2.8.2
python-editor==1.0.4
pytz==2021.1
PyYAML==5.4.1
pyzmq==22.1.0
requests==2.26.0
requests-unixsocket==0.2.0
rsa==4.7.2
scikit-allel==1.3.5
scikit-learn==0.24.2
scipy==1.7.0
seaborn==0.11.1
semver==2.13.0
Send2Trash==1.7.1
six==1.15.0
sklearn==0.0
sniffio==1.2.0
SQLAlchemy==1.4.21
statsmodels==0.12.2
stevedore==3.3.0
tabulate==0.8.9
tenacity==8.0.1
terminado==0.10.1
testpath==0.5.0
textwrap3==0.9.2
threadpoolctl==2.2.0
toolz==0.11.1
torch==1.9.0+cu102
torchvision==0.10.0+cu102
tornado==6.1
tqdm==4.61.2
traitlets==5.0.5
typing-extensions==3.10.0.0
urllib3==1.26.6
wcwidth==0.2.5
webencodings==0.5.1
websocket-client==1.1.0
widgetsnbextension==3.5.1
xarray==0.19.0